1.将刷新方式改为——下载到数据以后进行才删除原始数据。2.优化@界面的翻转显示，将载入数据延迟到翻转完成后进行。

+++ SocialFusion/Class/StatusDetailController.m
@@ -22,13 +22,23 @@
 #import "RenrenUser.h"
 #import "WeiboUser.h"
 
-
 @implementation StatusDetailController
 
-@synthesize feedData=_feedData;
+@synthesize feedData = _feedData;
 
+- (void)dealloc {
+    [_pageLine removeFromSuperview];
+    [_pageLine release];
+    [self.feedData release];
+    //[_feedStatusCel release];
+    [super dealloc];
+}
 
-
+- (void)viewDidUnload
+{
+    
+    [super viewDidUnload];
+}
 
 - (void)showBigImage
 {
@@ -42,47 +52,40 @@
 
 - (BOOL)webView:(UIWebView *)webView shouldStartLoadWithRequest:(NSURLRequest *)request navigationType:(UIWebViewNavigationType)navigationType
 {
-    
-    
-    NSString* tempString=[NSString stringWithFormat:@"%@",[request URL]];
+    NSString* tempString = [NSString stringWithFormat:@"%@",[request URL]];
     //NSLog(@"%@",tempString);
     
-    NSString* commandString=[tempString substringFromIndex:7];
+    NSString* commandString = [tempString substringFromIndex:7];
     if ([commandString isEqualToString:@"showimage"])
     {
         [self showBigImage];
         return NO;
     }
-
+    
     return YES;
 }
 
-
-
-
 - (void)clearData
 {
+    if(!_clearDataFlag)
+        return;
+    _clearDataFlag = NO;
     
     _noAnimationFlag = YES;
-  //  NSLog(@"%@",self.feedData.comments);
+    //  NSLog(@"%@",self.feedData.comments);
     [self.feedData removeComments:self.feedData.comments];
     
     [StatusCommentData deleteAllObjectsInManagedObjectContext:self.managedObjectContext];
- 
+    
 }
-
-
-
-
-
 
 - (void)setFixedInfo
 {
     
-    _pageLine=[[UIImageView alloc] initWithImage:[UIImage imageNamed:@"page_line.png"]];
-    _pageLine.frame=CGRectMake(305, 0, 1, 350);
+    _pageLine = [[UIImageView alloc] initWithImage:[UIImage imageNamed:@"page_line.png"]];
+    _pageLine.frame = CGRectMake(305, 0, 1, 350);
     
-    _nameLabel.text=[_feedData getFeedName];
+    _nameLabel.text = [_feedData getFeedName];
     NSData *imageData = nil;
     if([Image imageWithURL:_feedData.owner_Head inManagedObjectContext:self.managedObjectContext]) {
         imageData = [Image imageWithURL:_feedData.owner_Head inManagedObjectContext:self.managedObjectContext].imageData.data;
@@ -90,20 +93,20 @@
     if(imageData != nil) {
         _headImage.image = [UIImage imageWithData:imageData];
     }
-    _time.text=[CommonFunction getTimeBefore:_feedData.update_Time]; 
+    _time.text = [CommonFunction getTimeBefore:_feedData.update_Time]; 
     [(UIScrollView*)self.view setContentSize:CGSizeMake(self.view.frame.size.width*2,390)];
-    ((UIScrollView*)self.view).pagingEnabled=YES;
-    ((UIScrollView*)self.view).showsVerticalScrollIndicator=NO;
-    ((UIScrollView*)self.view).directionalLockEnabled=YES;
-    ((UIScrollView*)self.view).delegate=self;
+    ((UIScrollView*)self.view).pagingEnabled = YES;
+    ((UIScrollView*)self.view).showsVerticalScrollIndicator = NO;
+    ((UIScrollView*)self.view).directionalLockEnabled = YES;
+    ((UIScrollView*)self.view).delegate = self;
     [((UIScrollView*)self.view) addSubview:_pageLine];
-
+    
     self.tableView.frame = CGRectMake(306, 0, 306, 350);
     self.tableView.separatorStyle = UITableViewCellSeparatorStyleNone;
     [self.view addSubview:self.tableView];
-   
-    self.tableView.allowsSelection=NO;    
-    if ([_feedData.style intValue]==0)
+    
+    self.tableView.allowsSelection = NO;    
+    if ([_feedData.style intValue] == 0)
     {
         [_style setImage:[UIImage imageNamed:@"detail_renren.png"]];
     }
@@ -113,21 +116,15 @@
     }
     
     
+}
 
- 
+- (void)addOriStatus {
+    [self setFixedInfo];
+    [self loadMainView];
     
 }
 
-
-- (void)addOriStatus
-{
-    [self setFixedInfo];
-    [self loadMainView];
-
-}
-
-- (void) loadMainView
-{
+- (void) loadMainView {
     
 }
 
@@ -136,33 +133,29 @@
     self = [super initWithNibName:nibNameOrNil bundle:nibBundleOrNil];
     if (self) {
         // Custom initialization
-       // _commentArray=[[NSMutableArray alloc] init];
+        // _commentArray = [[NSMutableArray alloc] init];
         
     }
     return self;
 }
 
-
 - (void)viewWillAppear:(BOOL)animated
 {
     [super viewWillAppear:YES];
-      [self addOriStatus ];
+    [self addOriStatus ];
 }
+
 - (void)viewDidLoad
 {
     [super viewDidLoad];
-  
+    
 }
+
 - (void)viewDidAppear:(BOOL)animated
 {
     [super viewDidAppear:YES];
-    _pageNumber=0;
-   [self refresh];
-}
-- (void)viewDidUnload
-{
- 
-    [super viewDidUnload];
+    _pageNumber = 0;
+    [self refresh];
 }
 
 - (void)viewDidDisappear:(BOOL)animated
@@ -170,6 +163,7 @@
     //[_feedStatusCel removeFromSuperview];
     //[_feedStatusCel release];
 }
+
 - (void)didReceiveMemoryWarning
 {
     [super didReceiveMemoryWarning];
@@ -177,67 +171,46 @@
     [Image clearAllCacheInContext:self.managedObjectContext];
 }
 
-
-- (void)dealloc {
-    [_pageLine removeFromSuperview];
-    [_pageLine release];
-    [self.feedData release];
-    //[_feedStatusCel release];
-    [super dealloc];
-}
-
-
 - (void)configureRequest:(NSFetchRequest *)request
 {
     [request setEntity:[NSEntityDescription entityForName:@"StatusCommentData" inManagedObjectContext:self.managedObjectContext]];
     NSPredicate *predicate;
     NSSortDescriptor *sort;
-  
+    
     predicate = [NSPredicate predicateWithFormat:@"SELF IN %@",self.feedData.comments];
-   // NSLog(@"%@",self.feedData.comments);
+    // NSLog(@"%@",self.feedData.comments);
     sort = [[NSSortDescriptor alloc] initWithKey:@"update_Time" ascending:YES];
     [request setPredicate:predicate];
     NSArray *descriptors = [NSArray arrayWithObject:sort];
-
+    
     [request setSortDescriptors:descriptors]; 
     [sort release];
     request.fetchBatchSize = 5;
-
+    
 }
 
 #pragma mark - EGORefresh Method
-- (void)refresh {
-           [self hideLoadMoreDataButton];
-    [self clearData];
-
-    if ([_feedData getStyle]==0)
+- (void)refresh {    
+    [self hideLoadMoreDataButton];
+    _clearDataFlag = YES;
+    
+    if ([_feedData getStyle] == 0)
     {
-        _pageNumber=[_feedData getComment_Count]/10+1;
+        _pageNumber = [_feedData getComment_Count]/10+1;
     }
     else
     {
-        _pageNumber=1;
+        _pageNumber = 1;
     }
-    
-    
-
-    
-    if(_loading)
-        return;
-    _loading = YES;
-
     [self loadData];
-    
 }
 
-
 - (void)loadMoreData {
-    if(_loading)
+    if(_loadingFlag)
         return;
-    _loading = YES;
+    _loadingFlag = YES;
     
-    
-    if ([_feedData getStyle]==0)
+    if ([_feedData getStyle] == 0)
     {
         _pageNumber--;
     }
@@ -245,88 +218,89 @@
     {
         _pageNumber++;
     }
-    [self loadData]    ;
+    [self loadData];
 }
-
-
 
 - (void)ProcessRenrenData:(NSArray*)array
 {
     for(NSDictionary *dict in array) {
-        StatusCommentData* commentsData=[StatusCommentData insertNewComment:0 Dic:dict inManagedObjectContext:self.managedObjectContext];
-      
-    
+        StatusCommentData* commentsData = [StatusCommentData insertNewComment:0 Dic:dict inManagedObjectContext:self.managedObjectContext];
+        
+        
         if ([self.currentRenrenUser.userID isEqualToString:commentsData.actor_ID] )
         {
-            commentsData.actor_ID=[NSString stringWithString:@"self"];
+            commentsData.actor_ID = [NSString stringWithString:@"self"];
         }
         [_feedData addCommentsObject:commentsData];
-     
+        
         
     }
-    if (_pageNumber!=1)
+    if (_pageNumber != 1)
     {
-        _showMoreButton=YES;
+        _showMoreButton = YES;
     }
     else
     {
-        _showMoreButton=NO;
+        _showMoreButton = NO;
     }
     
-    _loading=NO;
+    _loadingFlag = NO;
     [self doneLoadingTableViewData];
-     [self.tableView reloadData];
+    [self.tableView reloadData];
 }
 
 - (void)ProcessWeiboData:(NSArray*)array
 {
     for(NSDictionary *dict in array) {
         
-        StatusCommentData* commentsData=[StatusCommentData insertNewComment:1 Dic:dict inManagedObjectContext:self.managedObjectContext];
+        StatusCommentData* commentsData = [StatusCommentData insertNewComment:1 Dic:dict inManagedObjectContext:self.managedObjectContext];
         
         if ([self.currentWeiboUser.userID isEqualToString:commentsData.actor_ID] )
         {
-            commentsData.actor_ID=[NSString stringWithString:@"self"];
+            commentsData.actor_ID = [NSString stringWithString:@"self"];
         }
         [_feedData addCommentsObject:commentsData]; 
     }
-    
-    
-    
     if ([_feedData.comments count]<[_feedData getComment_Count])
     {
-        _showMoreButton=YES;
+        _showMoreButton = YES;
     }
     else
     {
-        _showMoreButton=NO;
+        _showMoreButton = NO;
     }
-    _loading=NO;
+    _loadingFlag = NO;
     [self doneLoadingTableViewData];
     [self.tableView reloadData];
-
+    
 }
 
 - (void)loadData
 {
-    if ([_feedData getStyle]==0)
+    if(_loadingFlag)
+        return;
+    _loadingFlag = YES;
+    
+    if ([_feedData getStyle] == 0)
     {
         RenrenClient *renren = [RenrenClient client];
         [renren setCompletionBlock:^(RenrenClient *client) {
             if(!client.hasError) {
+                [self clearData];
                 NSArray *array = client.responseJSONObject;
                 [self ProcessRenrenData:array];
-              //[self.tableView reloadData];
+                //[self.tableView reloadData];
             }
-
+            
         }];
-            [renren getComments:[_feedData getActor_ID] status_ID:[_feedData getSource_ID] pageNumber:_pageNumber];
+        [renren getComments:[_feedData getActor_ID] status_ID:[_feedData getSource_ID] pageNumber:_pageNumber];
     }
     else
     {
         WeiboClient *weibo = [WeiboClient client];
         [weibo setCompletionBlock:^(WeiboClient *client) {
             if(!client.hasError) {
+                [self clearData];
                 NSArray *array = client.responseJSONObject;
                 [self ProcessWeiboData:array];
             }
@@ -335,40 +309,29 @@
     }
 }
 
-
-
-
-
-
-
 - (float) tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    
-
-        if (_showMoreButton==YES)
+    if (_showMoreButton == YES)
+    {
+        
+        if (indexPath.row == 0)
         {
-            
-            if (indexPath.row==0)
-            {
-                return 60;
-            }
-            else
-            {
-                NSIndexPath* index=[NSIndexPath indexPathForRow:indexPath.row-1 inSection:indexPath.section];
-                
-               // indexPath=[[indexPath indexPathByRemovingLastIndex] indexPathByRemovingLastIndex];
-                return [StatusCommentCell heightForCell:[self.fetchedResultsController objectAtIndexPath:index]];
-            }
+            return 60;
         }
         else
         {
-       //     indexPath=[indexPath indexPathByRemovingLastIndex];
-            return [StatusCommentCell heightForCell:[self.fetchedResultsController objectAtIndexPath:indexPath]];
+            NSIndexPath* index = [NSIndexPath indexPathForRow:indexPath.row-1 inSection:indexPath.section];
             
+            // indexPath = [[indexPath indexPathByRemovingLastIndex] indexPathByRemovingLastIndex];
+            return [StatusCommentCell heightForCell:[self.fetchedResultsController objectAtIndexPath:index]];
         }
-
-
-    
+    }
+    else
+    {
+        //     indexPath = [indexPath indexPathByRemovingLastIndex];
+        return [StatusCommentCell heightForCell:[self.fetchedResultsController objectAtIndexPath:indexPath]];
+        
+    }
     
 }
 
@@ -379,15 +342,11 @@
     return 1;
 }
 
-
-
-
-
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     
-    int number=[super tableView:tableView numberOfRowsInSection:section];
-    if (_showMoreButton==NO)
+    int number = [super tableView:tableView numberOfRowsInSection:section];
+    if (_showMoreButton == NO)
     {
         return number;
     }
@@ -400,33 +359,33 @@
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     static NSString *StatusComentCell = @"StatusCommentCell";
-     if (_showMoreButton==NO)
-     {
+    if (_showMoreButton == NO)
+    {
         StatusCommentCell* cell;
         cell = (StatusCommentCell *)[tableView dequeueReusableCellWithIdentifier:StatusComentCell];
         if (cell == nil) {
             [[NSBundle mainBundle] loadNibNamed:@"StatusCommentCell" owner:self options:nil];
             cell = _commentCel;
         }
-         
-         if (indexPath.row %2 ==0)
-         {
-             [cell configureCell:[self.fetchedResultsController objectAtIndexPath:indexPath] colorStyle:YES];
+        
+        if (indexPath.row %2 == 0)
+        {
+            [cell configureCell:[self.fetchedResultsController objectAtIndexPath:indexPath] colorStyle:YES];
             
-         }
-         else
-         {
-             [cell configureCell:[self.fetchedResultsController objectAtIndexPath:indexPath] colorStyle:NO];
-         }
-         
+        }
+        else
+        {
+            [cell configureCell:[self.fetchedResultsController objectAtIndexPath:indexPath] colorStyle:NO];
+        }
+        
         return cell;
-     }
+    }
     else
     {
-        if (indexPath.row==0)
+        if (indexPath.row == 0)
         {
-          UITableViewCell* cell=[[UITableViewCell alloc] initWithFrame:CGRectMake(0, 0, 306, 60)]; 
-           [cell.contentView addSubview:self.loadMoreDataButton];
+            UITableViewCell* cell = [[UITableViewCell alloc] initWithFrame:CGRectMake(0, 0, 306, 60)]; 
+            [cell.contentView addSubview:self.loadMoreDataButton];
             return cell;
         }
         else
@@ -439,25 +398,19 @@
                 cell = _commentCel;
             }
             
-            NSIndexPath* index=[NSIndexPath indexPathForRow:indexPath.row-1 inSection:indexPath.section];
+            NSIndexPath* index = [NSIndexPath indexPathForRow:indexPath.row-1 inSection:indexPath.section];
             
-            if (index.row %2 ==0)
+            if (index.row %2 == 0)
             {
-            [cell configureCell:[self.fetchedResultsController objectAtIndexPath:index] colorStyle:YES];
+                [cell configureCell:[self.fetchedResultsController objectAtIndexPath:index] colorStyle:YES];
             }
             else
             {
-            [cell configureCell:[self.fetchedResultsController objectAtIndexPath:index] colorStyle:NO];
+                [cell configureCell:[self.fetchedResultsController objectAtIndexPath:index] colorStyle:NO];
             }
             return cell;
         }
-    };  
-  
-    
+    }
 }
 
-
-
-
-
-    @end
+@end
